feat: add startLoadingAfter and stopLoadingAfter function for delaying
