+++ src/hooks/__tests__/use-loading.spec.ts
@@ -1,4 +1,4 @@
-import { useLoading } from '../useLoading'
+import { useLoading } from '../use-loading'
 import { renderHook } from '@testing-library/react'
 import { act } from 'react'
